Clean up of boundflux_flag==1
Cleaned up the calculation of the surface heat flux when using boundflux_flag which imitates a cooling plate. The old version was rotten, but the error only occured when melting occured at the surface, which ever happened in the standard testcases. New version uses the max_flax_plate as intended, I bumped the value high enough to not effect the standard testcase 1.