Changed temperature of falling snow
Thanks to a question by Caroline Meier I saw that falling snow always had a temperature of -1, which has now been changed so that falling snow has the two meter temperature as long as it is lower than -1. This has a small effect on the snow temperature, and it isn't really a bug fix as we do know the actual temperature of the falling snow to begin with.